Refactor tabs according to new approach
- URL collections:/id/<tab name>
- Detail page with tabs defaults to Details tab
- There's back button to List view
- Detail page lives in folder CollectionPage
- CollectionPage is parent component that calls API for all shared data and contains RoutedTabs
- All tabs have their own file named Collection<Tab name>
- All components live in coponents folder
- Interfaces are in Collection file

TODO:
- add Empty/Loading/Error states
