+++ frontend/hub/collections/CollectionPage/CollectionContents.tsx
@@ -1,0 +1,7 @@
+import { useTranslation } from 'react-i18next';
+import { DetailInfo } from '../../../../framework/components/DetailInfo';
+
+export function CollectionContents() {
+  const { t } = useTranslation();
+  return <DetailInfo title={t('This page is under construction')} />;
+}
