+++ frontend/hub/collections/CollectionPage/CollectionDistributions.tsx
@@ -1,0 +1,7 @@
+import { DetailInfo } from '../../../../framework/components/DetailInfo';
+import { useTranslation } from 'react-i18next';
+
+export function CollectionDistributions() {
+  const { t } = useTranslation();
+  return <DetailInfo title={t('This page is under construction')} />;
+}
